move stash-search files
