+++ renderer/src/web/overlay/OverlayWindow.vue
@@ -36,7 +36,7 @@ import { useI18n } from 'vue-i18n'
 import { Host } from '@/web/background/IPC'
 import { Widget, WidgetManager } from './interfaces'
 import WidgetTimer from './WidgetTimer.vue'
-import WidgetStashSearch from './WidgetStashSearch.vue'
+import WidgetStashSearch from '../stash-search/WidgetStashSearch.vue'
 import WidgetMenu from './WidgetMenu.vue'
 import PriceCheckWindow from '@/web/price-check/PriceCheckWindow.vue'
 import WidgetItemCheck from '@/web/item-check/WidgetItemCheck.vue'
