+++ renderer/src/web/stash-search/stash-search-editor.vue
@@ -33,8 +33,8 @@
 import { defineComponent } from 'vue'
 import { useI18n } from 'vue-i18n'
 import DndContainer from 'vuedraggable'
-import HotkeyInput from './HotkeyInput.vue'
-import { configProp, configModelValue } from './utils'
+import HotkeyInput from '../settings/HotkeyInput.vue'
+import { configProp, configModelValue } from '../settings/utils.js'
 import type { StashSearchWidget } from '@/web/overlay/interfaces'
 
 export default defineComponent({
